fix(core): 按 Copilot 审阅收紧研究测量
Gate: G2 Pass
Slice: core-runtime
Impact: core-api=none; data-format=none; adapter-api=none
Scope: 逐分量提升 residual-aware 差分、移出 Criterion 事件统计、将 10k 回归改为显式研究测试，并用重启后同一省电状态的命名 baseline 更新格式化表格
Validation: cargo +1.96.0 test --workspace --locked; cargo +1.96.0 clippy --workspace --all-targets --locked -- -D warnings; cargo-deny check advisories bans licenses sources; cargo +1.96.0 fmt --all -- --check; xtask format-md-tables --check; 10k/100k differential 与 Criterion 矩阵通过
Docs: updated

Refs: #140

+++ crates/laneflow-core/tests/support/numeric_precision_research.rs
@@ -244,8 +244,8 @@ macro_rules! residual_aware_progress_methods {
             follower_residual: f32,
         ) -> Self::Compute {
             Self::Compute::from_f64(
-                f64::from(leader_progress - follower_progress)
-                    - f64::from(leader_residual - follower_residual),
+                (f64::from(leader_progress) - f64::from(follower_progress))
+                    - (f64::from(leader_residual) - f64::from(follower_residual)),
             )
         }
 
@@ -254,7 +254,9 @@ macro_rules! residual_aware_progress_methods {
             progress: f32,
             residual: f32,
         ) -> Self::Compute {
-            Self::Compute::from_f64(f64::from(edge_length - progress) + f64::from(residual))
+            Self::Compute::from_f64(
+                (f64::from(edge_length) - f64::from(progress)) + f64::from(residual),
+            )
         }
 
         fn progress_reaches_boundary(
